Add files via upload
In this project, we explore a general theory of multiple-phase competition for classical Heisenberg
spins with anisotropic exchange interactions on a square-kagome lattice. We started with determining
the generic nearest-neighbour Hamiltonian respecting the D4 point-group symmetry of the lattice.
The structure of the coupling matrices in this Hamiltonian indicates that the system supports XYZ
model with out-of-plane Dzyaloshinskii-Moriya (DM) interactions. Then we employed represention
theory to determine the orthonormal basis spin configurations and the corresponding order parameters
in order to diagonalize the Hamiltonian, and hence dertermine the ground state phases for this generic
model. We performed a Monte Carlo simulation for the specific case of XXZ model with out-of-
plane Dzyaloshinskii-Moriya couplings and obtained a rich phase diagram of the ground state with
five distinct classical ground-state spin configurations. We are yet to study the magnetic excitation
spectra of the ground state phases and analyze their topological properties and transport signatures.